# [Performance] Significantly Improve Client Network Resends