--- a/common/net/daybreak_connection.h
+++ b/common/net/daybreak_connection.h
@@ -181,6 +181,10 @@ namespace EQ
 			Timestamp m_close_time;
 			double m_outgoing_budget;
 
+			// resend tracking
+			size_t m_resend_packets_sent = 0;
+			size_t m_resend_bytes_sent = 0;
+
 			struct DaybreakSentPacket
 			{
 				DynamicPacket packet;
@@ -317,6 +321,10 @@ namespace EQ
 		private:
 			void Attach(uv_loop_t *loop);
 			void Detach();
+
+			// resend
+			Timestamp m_last_resend_check;
+			int m_resend_check_interval = 50;
 
 			EQ::Random m_rand;
 			uv_timer_t m_timer;