### Description

This PR addresses several client network issues reported by players, aiming to significantly improve connection reliability and speed. These changes mitigate common disconnect scenarios, especially during high-stress network events, such as heavy raid traffic or zoning with large inventories.

---

### Player-Perceived Issues

While these improvements greatly enhance connection stability, it’s important to note that extreme cases (e.g., very poor connections) may still face limitations due to factors outside the server's control.

Common issues addressed:
- Players disconnect during heavy network traffic (e.g., raid burns).
- Players experience "random" disconnects while zoning.
- Players with poor connections disconnecting while doing seemingly nothing.
- Players with poor connections disconnecting while trying to enter world with large inventories.

---

### Root Cause Analysis

These issues are deeply rooted in network-layer behavior. When either the server or client connection falls behind, packets are queued for resending. This process introduces several challenges:

#### Resend Storms
A typical scenario involves a player zoning in with a large inventory on a limited bandwidth connection. The server queues a massive number of packets for transmission (e.g., up to **1MB pre-compression**). If the client cannot keep up:
- The server aggressively resends packets every 16ms (~62.5 times per second).
- This can result in a compressed data load of 18.75MB/s, overwhelming the client and causing disconnects.
- Players with large inventories or poor connections are particularly affected, often entering a perpetual resend state that exacerbates disconnect issues.

---

### Solutions

The following changes are implemented to resolve these issues:

1. **Resend Window Optimization**
   - Limit resend attempts to a maximum of **300 packets or 140KB of data per window** (based on observed client processing capabilities).
   - Significantly reduces expensive server memory allocation and network bandwidth usage.

2. **Reduced Resend Frequency**
   - Resend checks are now performed every 50ms instead of 16ms, decreasing the rate of packet resends.

3. **Targeted Resend Checks**
   - Resend attempts now prioritize the first overdue packet, rather than iterating through all queued packets.

4. **Early Exit on Empty Queues**
   - Resend processing halts immediately when no packets remain in the queue, preventing unnecessary overhead.

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] Performance improvement

# Testing

Stress testing bags. A client with 36 x 100 slot bags with 100 items each = (3600 items) would take upwards of 30-40 seconds to load if not more if not entirely disconnect originally. With these changes the client loads in about 5-6 seconds.

![image](https://github.com/user-attachments/assets/3bb6fd0f-b67d-4923-b995-116bb244ed73)

https://github.com/user-attachments/assets/42211213-09e2-45d2-b065-dad848819d66

# Checklist

- [x] I have tested my changes
- [x] I have performed a self-review of my code. Ensuring variables, functions and methods are named in a human-readable way, comments are added only where naming of variables, functions and methods can't give enough context.
- [x] I own the changes of my code and take responsibility for the potential issues that occur